Refactor mint-diet mock to comply with review feedback
## Changes

### File Splitting (≤200 lines)
- Split plans.tsx (403 lines) into:
  - plans-list.tsx (73 lines) - plan listing and creation
  - plans-detail.tsx (99 lines) - plan details, edit, delete
  - plans-meals-items.tsx (122 lines) - meal plan items
  - plans-meals-ingredients.tsx (90 lines) - ingredients

- Split log.tsx (308 lines) into:
  - log-views.tsx (64 lines) - HTML page routes
  - log-entries-create.tsx (108 lines) - create entries
  - log-entries-update.tsx (119 lines) - update entries
  - log-entries-delete.tsx (50 lines) - delete entries

### Bug Fixes
- Fix routes/index.tsx re-exports
- helpers.tsx: jsonError now supports 400/404/500 status codes
- Add /** @jsxImportSource hono/jsx */ pragma to all TSX files
- Update tests to import from log-shared instead of log

### Improvements
- index.tsx: Add healthResponse, remove type cast
- types.ts: MintDietApp is now an alias for OpenAPIApp

### Verification
- All tests pass (16 pass)
- TypeScript build clean
- OpenAPI generation works
