+++ mock-platform/mocks/mint-diet/src/components/Layout.tsx
@@ -1,3 +1,4 @@
+/** @jsxImportSource hono/jsx */
 import type { FC } from "hono/jsx";
 import { CSS } from "../styles";
