+++ mock-platform/mocks/mint-diet/src/routes/admin.tsx
@@ -1,8 +1,7 @@
 import type { Database } from "bun:sqlite";
 import { createRoute } from "mock-lib";
-import { Layout } from "../components";
-import { isResponse, runDbMutation } from "./helpers";
-import { HtmlResponse, RedirectResponse } from "./schemas";
+import { isResponse, jsonError, runDbMutation } from "./helpers";
+import { RedirectResponse } from "./schemas";
 import type { MintDietApp, RouteDeps } from "./types";
 
 export function resetMutableTables(d: Database): void {
@@ -34,14 +33,28 @@ export function registerAdminRoutes(app: MintDietApp, { getDatabase }: RouteDeps
     summary: "Reset mutable diet data",
     responses: {
       303: RedirectResponse,
-      404: HtmlResponse,
-      500: HtmlResponse,
+      404: {
+        description: "Not found - admin not enabled",
+        content: {
+          "application/json": {
+            schema: { type: "object", properties: { success: { type: "boolean" }, message: { type: "string" } } },
+          },
+        },
+      },
+      500: {
+        description: "Server error",
+        content: {
+          "application/json": {
+            schema: { type: "object", properties: { success: { type: "boolean" }, message: { type: "string" } } },
+          },
+        },
+      },
     },
   });
 
   app.openApiRoute(resetRoute, async (c) => {
     if (!process.env.MOCK_ADMIN || process.env.MOCK_ADMIN !== "1") {
-      return c.html(<Layout title="Not Found"><p>Not found</p></Layout>, 404);
+      return jsonError(c, "Not found", 404);
     }
 
     const d = getDatabase();
